slow swipe added
